Fix -Werrors regarding va_arg types

+++ snmplib/snprintf.c
@@ -243,7 +243,7 @@ dopr(char *buffer, size_t maxlen, const char *format, va_list args)
             case 'd':
             case 'i':
                 if (cflags == DP_C_SHORT)
-                    value = va_arg(args, short int);
+                    value = va_arg(args, long);
                 else if (cflags == DP_C_LONG)
                     value = va_arg(args, long int);
                 else
@@ -254,7 +254,7 @@ dopr(char *buffer, size_t maxlen, const char *format, va_list args)
             case 'o':
                 flags |= DP_F_UNSIGNED;
                 if (cflags == DP_C_SHORT)
-                    value = va_arg(args, unsigned short int);
+                    value = va_arg(args, unsigned long);
                 else if (cflags == DP_C_LONG)
                     value = va_arg(args, unsigned long int);
                 else
@@ -265,7 +265,7 @@ dopr(char *buffer, size_t maxlen, const char *format, va_list args)
             case 'u':
                 flags |= DP_F_UNSIGNED;
                 if (cflags == DP_C_SHORT)
-                    value = va_arg(args, unsigned short int);
+                    value = va_arg(args, unsigned long);
                 else if (cflags == DP_C_LONG)
                     value = va_arg(args, unsigned long int);
                 else
@@ -278,7 +278,7 @@ dopr(char *buffer, size_t maxlen, const char *format, va_list args)
             case 'x':
                 flags |= DP_F_UNSIGNED;
                 if (cflags == DP_C_SHORT)
-                    value = va_arg(args, unsigned short int);
+                    value = va_arg(args, unsigned long);
                 else if (cflags == DP_C_LONG)
                     value = va_arg(args, unsigned long int);
                 else
